% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileControlTest.R
\name{quantileControlTest}
\alias{quantileControlTest}
\title{Find standard error for survival quantile}
\usage{
quantileControlTest(timevar1, censor1, timevar2, censor2, q = 0.5, B = 1000,
  seed = 1234, plots = FALSE)
}
\arguments{
\item{timevar1}{Vector of observed survival times for sample 1 (control).}

\item{censor1}{Vector of censoring indicators for sample 1 (1 = uncensored, 0 = censored).}

\item{timevar2}{Vector of observed survival times for sample 2 (treatment).}

\item{censor2}{Vector of censoring indicators for sample 2 (1 = uncensored, 0 = censored).}

\item{q}{Quantile of interest (Default is median).}

\item{B}{Number of bootstrap samples.}

\item{seed}{Seed number (for reproducibility).}

\item{plots}{Logical. TRUE to show Kaplan-Meier plot}
}
\value{
Returns quantile estimate, bootstrapped standard error, test statistic, and two-sided p-value.
}
\description{
Find standard error for survival quantile
}
\details{
It is important to note the possiblilty that the estimated quantile may not be estimable in our bootstrap samples. In such cases
the largest observed survival time will be considered as an estimate for the quantile.
}
\examples{
#Reference: Survival Analysis Techniques for Censored and Truncated Data.
#Klein and Moeschberger (1997) Springer.
#Data: Chapter 7.6 Example 7.9 (p. 211)
library(controlTest)
t1 <- c(1, 63, 105, 129, 182, 216, 250, 262, 301, 301,
       342, 354, 356, 358, 380, 383, 383, 338, 394, 408, 460, 489,
       499, 523, 524, 535, 562, 569, 675, 676, 748, 778, 786, 797,
       955, 968, 1000, 1245, 1271, 1420, 1551, 1694, 2363, 2754, 2950)
t2 <- c(17, 42, 44, 48, 60, 72, 74, 95, 103, 108, 122, 144, 167, 170,
       183, 185, 193, 195, 197, 208, 234, 235, 254, 307, 315, 401, 445,
       464, 484, 528, 542, 547, 577, 580, 795, 855, 1366, 1577, 2060,
       2412, 2486, 2796, 2802, 2934, 2988)
c1 <- c(rep(1, 43), 0, 0)
c2 <- c(rep(1, 39), rep(0, 6))
quantileControlTest(t1, c1, t2, c2, q = 0.5, B = 500)

}
\references{
Li, G., Tiwari, R.C., and Wells, M. (1996). "Quantile Comparison Functions in Two-Sample Problems: With Applications to Comparisons of Diagnostic Markers." Journal of the American Statistical Association, 91, 689-698.

Chakraborti, S., and Mukerjee, R. (1989), "A Confidence Interval for a Measure Associated With the Comparison of a Treatment With a Control," South African Statistical Journal, 23, 219-230.

Gastwirth, J. L., and Wang, J. L. (1988), "Control Percentile Test for Censored Data," Journal of Statistical Planning and Inference, 18, 267-276.
}
